Finally, something to code on
Credit: GitHub Copilot. The prompts were "Screenshot 2025-03-26 172401.png
Please write a .php file that resembles this image. The image depicts a user input field labelled, "Enter Instructor Name." The user is able to input the instructor name "Luis Jaimes." Once the input is typed in, a button can be pressed that reads "Search Courses." Pressing this button reads "Example Instructor's Courses" from an .sql file and displays it for the user in a table below. All the courses listed in the table are assumed to be taught by Luis Jaimes.", "I would like to make this my default webpage. Please encase this php logic inside of a file called "index.html"", "On line 10 of index.php, do not include "Luis Jaimes" as the value. Leave the box blank" and "This php code will execute sql logic based on user input. Please recode this so it's invulnerable to SQL injection attacks."